Split analyzer module into base/posterior/empirical + factory
- Break the monolithic analyzer.py into analyzer_base, posterior_analyzer, and
  empirical_analyzer modules along the neural/empirical family seam.
- Move the make_analyzer dispatch into factory.py, re-exported from the package
  __init__; lazy concrete imports keep package import free of torch/sbi so the
  eval-time episode_writer path stays light.
