+++ isaaclab_arena/analysis/sensitivity/plotting.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from isaaclab_arena.analysis.sensitivity.analyzer import BaseAnalyzer
+    from isaaclab_arena.analysis.sensitivity.analyzer_base import BaseAnalyzer
     from isaaclab_arena.analysis.sensitivity.dataset import FactorSpec
